move content folder to the pagecontent
