--- FCmsManager/Areas/FCmsManager/Controllers/ContenteditorController.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Linq;
-using Microsoft.AspNetCore.Mvc;
-using FCms.Content;
-using FCmsManager.ViewModel;
-using Microsoft.AspNetCore.Authorization;
-
-namespace FCmsManager.Areas.FCmsManager.Controllers
-{
-    [Area("fcmsmanager")]
-    [Authorize(AuthenticationSchemes = "fcms")]
-    public class ContenteditorController: Controller
-    {
-
-    }
-}
