+++ FCmsManager/Areas/FCmsManager/Views/Content/Index.cshtml
@@ -11,7 +11,7 @@
 
             <div class="cmscontrols">
                 <a href="~/fcmsmanager/content/show?repositoryid=@repository.Id">Edit</a> |
-                <a href="~/fcmsmanager/definition?repositoryid=@repository.Id">Page structure</a> |
+                <a href="~/fcmsmanager/definition?repositoryid=@repository.Id">Content structure</a> |
                 <a href="~/fcmsmanager/repository/delete?repositoryid=@repository.Id">Delete</a>
             </div>
         </div>
