+++ FCmsManager/Areas/FCmsManager/Views/Definition/Index.cshtml
@@ -11,7 +11,7 @@
 
     <div class="cmscontrols">
         <a href="/fcmsmanager/definition/edit?repositoryid=@Model.Id&id=@definition.DefinitionId">Definition</a> |
-        <a href="/fcmsmanager/content/list?repositoryid=@Model.Id&definitionid=@definition.DefinitionId">List content</a> |
+        <a href="/fcmsmanager/pagecontent/list?repositoryid=@Model.Id&definitionid=@definition.DefinitionId">List content</a> |
         <a href="/fcmsmanager/definition/deletedefinition?repositoryid=@Model.Id&id=@definition.DefinitionId">Delete</a>
     </div>
 </div>}
